Add Connection.java Add JSONParser.java
Sınıfları ekleyerek
