+++ app/src/main/java/com/yildizozan/yurtbasi/MainActivity.java
@@ -4,7 +4,7 @@ import android.app.Activity;
 import android.os.Bundle;
 
 /**
- * Created by Ozan on 8/10/2016.
+ * Created by Ozan Yıldız on 2016/8/5.
  */
 
 public class MainActivity extends Activity {
