+++ app/src/main/java/com/yildizozan/yurtbasi/PasswordVerifyActivity.java
@@ -4,7 +4,7 @@ import android.app.Activity;
 import android.os.Bundle;
 
 /**
- * Created by Ozan on 8/10/2016.
+ * Created by Ozan Yıldız on 2016/8/5.
  */
 
 public class PasswordVerifyActivity extends Activity {
@@ -13,7 +13,6 @@ public class PasswordVerifyActivity extends Activity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_passwordverify);
-
 
     }
 }
